# plan_vision_backend/api-docs.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/projects": {
      "post": {
        "operationId": "ProjectController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateProjectDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "The project has been successfully created."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Create a new project.",
        "tags": [
          "Projects"
        ]
      },
      "get": {
        "operationId": "ProjectController_findAll",
        "parameters": [],
        "responses": {
          "200": {
            "description": "List of projects."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "List all projects belonging to the current user.",
        "tags": [
          "Projects"
        ]
      }
    },
    "/projects/{id}": {
      "get": {
        "operationId": "ProjectController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "The project details."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          },
          "404": {
            "description": "Project not found or does not belong to user."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Get a specific project by ID.",
        "tags": [
          "Projects"
        ]
      },
      "patch": {
        "operationId": "ProjectController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateProjectDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "The updated project."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          },
          "404": {
            "description": "Project not found or does not belong to user."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Update a project.",
        "tags": [
          "Projects"
        ]
      },
      "delete": {
        "operationId": "ProjectController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "The project has been successfully deleted."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          },
          "404": {
            "description": "Project not found or does not belong to user."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Delete a project.",
        "tags": [
          "Projects"
        ]
      }
    },
...
    "/render-configs": {
      "post": {
        "operationId": "RenderConfigController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateRenderConfigDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Render config created."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Create a render job with specified settings.",
        "tags": [
          "Render Configs"
        ]
      },
      "get": {
        "operationId": "RenderConfigController_findAll",
        "parameters": [
          {
            "name": "projectId",
            "required": true,
            "in": "query",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of render configs."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "List render configs for a project (provide projectId).",
        "tags": [
          "Render Configs"
        ]
      }
    },
    "/render-configs/{id}": {
      "get": {
        "operationId": "RenderConfigController_findOne",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Render config details."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          },
          "404": {
            "description": "Render config not found or access denied."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Get a single render config by ID.",
        "tags": [
          "Render Configs"
        ]
      },
      "patch": {
        "operationId": "RenderConfigController_update",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateRenderConfigDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Render config updated."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          },
          "404": {
            "description": "Render config not found or access denied."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Update metadata, instructions, or parameters.",
        "tags": [
          "Render Configs"
        ]
      },
      "delete": {
        "operationId": "RenderConfigController_remove",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Render config deleted."
          },
          "401": {
            "description": "Unauthorized. Session is required."
          },
          "404": {
            "description": "Render config not found or access denied."
          }
        },
        "security": [
          {
            "access-token": []
          },
          {
            "cookie": []
          }
        ],
        "summary": "Delete a render config.",
        "tags": [
          "Render Configs"
        ]
      }
    }
  },
  "info": {
    "title": "Plan Vision API",
    "description": "REST API for Plan Vision backend (users, auth, providers).",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "securitySchemes": {
      "cookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "auth_session",
        "description": "Better Auth session cookie"
      },
      "access-token": {
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "type": "http",
        "description": "Optional Better Auth access token"
      }
    },
    "schemas": {
...
      "CreateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Project name chosen by the user.",
            "maxLength": 255,
            "example": "Downtown Loft Floor Plan"
          }
        },
        "required": [
          "name"
        ]
      },
      "UpdateProjectDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "description": "Project name chosen by the user.",
            "maxLength": 255,
            "example": "Downtown Loft Floor Plan"
          }
        }
      },
...
      "CreateRenderConfigDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "number",
            "description": "Project ID this job belongs to."
          },
          "inputImageId": {
            "type": "number",
            "description": "Input image ID for the render."
          },
          "imageTypeId": {
            "type": "number",
            "description": "Image type ID describing the source."
          },
          "styleId": {
            "type": "number",
            "description": "Optional style ID chosen by the user."
          },
          "customInstructions": {
            "type": "string",
            "description": "Additional custom instructions."
          },
          "colorPrimaryHex": {
            "type": "string",
            "description": "Primary hex color, e.g., \"#3B82F6\"."
          },
          "colorSecondaryHex": {
            "type": "string",
            "description": "Secondary hex color."
          },
          "colorNeutralHex": {
            "type": "string",
            "description": "Neutral hex color."
          },
          "perspectiveAngle": {
            "type": "number",
            "description": "Perspective angle in degrees."
          },
          "perspectiveX": {
            "type": "number",
            "description": "Perspective X coordinate."
          },
          "perspectiveY": {
            "type": "number",
            "description": "Perspective Y coordinate."
          }
        },
        "required": [
          "projectId",
          "inputImageId",
          "imageTypeId"
        ]
      },
      "UpdateRenderConfigDto": {
        "type": "object",
        "properties": {
          "projectId": {
            "type": "number",
            "description": "Project ID this job belongs to."
          },
          "inputImageId": {
            "type": "number",
            "description": "Input image ID for the render."
          },
          "imageTypeId": {
            "type": "number",
            "description": "Image type ID describing the source."
          },
          "styleId": {
            "type": "number",
            "description": "Optional style ID chosen by the user."
          },
          "customInstructions": {
            "type": "string",
            "description": "Additional custom instructions."
          },
          "colorPrimaryHex": {
            "type": "string",
            "description": "Primary hex color, e.g., \"#3B82F6\"."
          },
          "colorSecondaryHex": {
            "type": "string",
            "description": "Secondary hex color."
          },
          "colorNeutralHex": {
            "type": "string",
            "description": "Neutral hex color."
          },
          "perspectiveAngle": {
            "type": "number",
            "description": "Perspective angle in degrees."
          },
          "perspectiveX": {
            "type": "number",
            "description": "Perspective X coordinate."
          },
          "perspectiveY": {
            "type": "number",
            "description": "Perspective Y coordinate."
          }
        }
      }
    }
  }
}